Isolate `pr list` command
