+++ command/pr.go
@@ -2,17 +2,11 @@ package command
 
 import (
 	"fmt"
-	"strconv"
 
 	"github.com/MakeNowJust/heredoc"
 	"github.com/cli/cli/api"
-	"github.com/cli/cli/internal/ghrepo"
-	"github.com/cli/cli/pkg/cmd/pr/shared"
-	"github.com/cli/cli/pkg/cmdutil"
-	"github.com/cli/cli/pkg/text"
 	"github.com/cli/cli/utils"
 	"github.com/spf13/cobra"
-	"github.com/spf13/pflag"
 )
 
 func init() {
@@ -22,14 +16,6 @@ func init() {
 	prCmd.AddCommand(prCloseCmd)
 	prCmd.AddCommand(prReopenCmd)
 	prCmd.AddCommand(prReadyCmd)
-
-	prCmd.AddCommand(prListCmd)
-	prListCmd.Flags().BoolP("web", "w", false, "Open the browser to list the pull request(s)")
-	prListCmd.Flags().IntP("limit", "L", 30, "Maximum number of items to fetch")
-	prListCmd.Flags().StringP("state", "s", "open", "Filter by state: {open|closed|merged|all}")
-	prListCmd.Flags().StringP("base", "B", "", "Filter by base branch")
-	prListCmd.Flags().StringSliceP("label", "l", nil, "Filter by labels")
-	prListCmd.Flags().StringP("assignee", "a", "", "Filter by assignee")
 }
 
 var prCmd = &cobra.Command{
@@ -48,18 +34,6 @@ var prCmd = &cobra.Command{
 - by URL, e.g. "https://github.com/OWNER/REPO/pull/123"; or
 - by the name of its head branch, e.g. "patch-1" or "OWNER:patch-1".`},
 }
-var prListCmd = &cobra.Command{
-	Use:   "list",
-	Short: "List and filter pull requests in this repository",
-	Args:  cmdutil.NoArgsQuoteReminder,
-	Example: heredoc.Doc(`
-	$ gh pr list --limit 999
-	$ gh pr list --state closed
-	$ gh pr list --label "priority 1" --label "bug"
-	$ gh pr list --web
-	`),
-	RunE: prList,
-}
 var prCloseCmd = &cobra.Command{
 	Use:   "close {<number> | <url> | <branch>}",
 	Short: "Close a pull request",
@@ -77,131 +51,6 @@ var prReadyCmd = &cobra.Command{
 	Short: "Mark a pull request as ready for review",
 	Args:  cobra.MaximumNArgs(1),
 	RunE:  prReady,
-}
-
-func prList(cmd *cobra.Command, args []string) error {
-	ctx := contextForCommand(cmd)
-	apiClient, err := apiClientForContext(ctx)
-	if err != nil {
-		return err
-	}
-
-	baseRepo, err := determineBaseRepo(apiClient, cmd, ctx)
-	if err != nil {
-		return err
-	}
-
-	web, err := cmd.Flags().GetBool("web")
-	if err != nil {
-		return err
-	}
-
-	limit, err := cmd.Flags().GetInt("limit")
-	if err != nil {
-		return err
-	}
-	if limit <= 0 {
-		return fmt.Errorf("invalid limit: %v", limit)
-	}
-
-	state, err := cmd.Flags().GetString("state")
-	if err != nil {
-		return err
-	}
-	baseBranch, err := cmd.Flags().GetString("base")
-	if err != nil {
-		return err
-	}
-	labels, err := cmd.Flags().GetStringSlice("label")
-	if err != nil {
-		return err
-	}
-	assignee, err := cmd.Flags().GetString("assignee")
-	if err != nil {
-		return err
-	}
-
-	if web {
-		prListURL := ghrepo.GenerateRepoURL(baseRepo, "pulls")
-		openURL, err := listURLWithQuery(prListURL, filterOptions{
-			entity:     "pr",
-			state:      state,
-			assignee:   assignee,
-			labels:     labels,
-			baseBranch: baseBranch,
-		})
-		if err != nil {
-			return err
-		}
-		fmt.Fprintf(cmd.ErrOrStderr(), "Opening %s in your browser.\n", utils.DisplayURL(openURL))
-		return utils.OpenInBrowser(openURL)
-	}
-
-	var graphqlState []string
-	switch state {
-	case "open":
-		graphqlState = []string{"OPEN"}
-	case "closed":
-		graphqlState = []string{"CLOSED", "MERGED"}
-	case "merged":
-		graphqlState = []string{"MERGED"}
-	case "all":
-		graphqlState = []string{"OPEN", "CLOSED", "MERGED"}
-	default:
-		return fmt.Errorf("invalid state: %s", state)
-	}
-
-	params := map[string]interface{}{
-		"state": graphqlState,
-	}
-	if len(labels) > 0 {
-		params["labels"] = labels
-	}
-	if baseBranch != "" {
-		params["baseBranch"] = baseBranch
-	}
-	if assignee != "" {
-		params["assignee"] = assignee
-	}
-
-	listResult, err := api.PullRequestList(apiClient, baseRepo, params, limit)
-	if err != nil {
-		return err
-	}
-
-	hasFilters := false
-	cmd.Flags().Visit(func(f *pflag.Flag) {
-		switch f.Name {
-		case "state", "label", "base", "assignee":
-			hasFilters = true
-		}
-	})
-
-	title := listHeader(ghrepo.FullName(baseRepo), "pull request", len(listResult.PullRequests), listResult.TotalCount, hasFilters)
-	if connectedToTerminal(cmd) {
-		fmt.Fprintf(colorableErr(cmd), "\n%s\n\n", title)
-	}
-
-	table := utils.NewTablePrinter(cmd.OutOrStdout())
-	for _, pr := range listResult.PullRequests {
-		prNum := strconv.Itoa(pr.Number)
-		if table.IsTTY() {
-			prNum = "#" + prNum
-		}
-		table.AddField(prNum, nil, shared.ColorFuncForPR(pr))
-		table.AddField(text.ReplaceExcessiveWhitespace(pr.Title), nil, nil)
-		table.AddField(pr.HeadLabel(), nil, utils.Cyan)
-		if !table.IsTTY() {
-			table.AddField(prStateWithDraft(&pr), nil, nil)
-		}
-		table.EndRow()
-	}
-	err = table.Render()
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
 
 func prClose(cmd *cobra.Command, args []string) error {
@@ -264,14 +113,6 @@ func prReopen(cmd *cobra.Command, args []string) error {
 	fmt.Fprintf(colorableErr(cmd), "%s Reopened pull request #%d (%s)\n", utils.Green("✔"), pr.Number, pr.Title)
 
 	return nil
-}
-
-func prStateWithDraft(pr *api.PullRequest) string {
-	if pr.IsDraft && pr.State == "OPEN" {
-		return "DRAFT"
-	}
-
-	return pr.State
 }
 
 func prReady(cmd *cobra.Command, args []string) error {
